// src/pages/school-landing/types.ts
import type { SchoolRecord } from '../../utils/subdomain';

export type { SchoolRecord };

export interface SchoolColors {
  primary: string;
  secondary: string;
  primaryRgb: string;
  btn: React.CSSProperties;
  iconBg: React.CSSProperties;
  text: React.CSSProperties;
  lightBg: React.CSSProperties;
  ctaBg: React.CSSProperties;
  borderColor: React.CSSProperties;
}